+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Bugfixes
 
 - lint cli and webpack-packages
 - fixed typo in documentation
+- cli
+  - Fixed issue with child_process.spawn on windows by using cross-spawn. (https://stackoverflow.com/questions/37459717/error-spawn-enoent-on-windows) #11 #12
+  - init project ignores .git directory
 
 ---